Fix excessive CPU usage when a station only sends data to PLC
A receive thread used to be created even when a station was only used
for sending (inSize=0). This thread looped constantly as there was no
data to receive, and consequently used huge amounts of CPU.

This was fixed by adding a check for inSize parameter and adding
(re)connection functionality also to the sending thread.