Remove base.dbd from s7plc.dbd
The support library and test IOC were not properly separated: base.dbd
was included in the support library's dbd instead of the test app's dbd.
Consequently, any application that used the s7plc module had to include
s7plc.dbd *first*, and *not* include base.dbd itself. This is contrary
to the usual practice.